+++ packages/spacecat-shared-data-access/src/index.d.ts
@@ -315,6 +315,27 @@ export interface Organization {
   getConfig: () => Config;
 }
 
+export interface Configuration {
+  /**
+   * Retrieves the configuration version.
+   * @returns {string} The configuration version.
+   */
+  getVersion: () => string;
+
+  /**
+   * Retrieves the queues configuration.
+   * @returns {object} The queues configuration.
+   */
+  getQueues: () => object;
+
+  /**
+   * Retrieves the jobs configuration.
+   * @returns {Array} The jobs configurations.
+   */
+  getJobs: () => Array<object>;
+
+}
+
 export interface DataAccess {
   getAuditForSite: (
     sitedId: string,
@@ -407,6 +428,10 @@ export interface DataAccess {
   upsertSiteCandidate: (siteCandidateDate: object) => Promise<SiteCandidate>;
   siteCandidateExists: (baseURL: string) => Promise<boolean>;
   updateSiteCandidate: (siteCandidate: SiteCandidate) => Promise<SiteCandidate>;
+
+  // configuration functions
+  getConfiguration: () => Promise<Configuration>
+  getConfigurationByVersion: (version: string) => Promise<Configuration>
 }
 
 interface DataAccessConfig {
@@ -415,6 +440,7 @@ interface DataAccessConfig {
   tableNameOrganizations: string,
   tableNameSites: string;
   tableNameSiteCandidates: string;
+  tableNameConfigurations: string;
   indexNameAllSites: string;
   indexNameAllSitesOrganizations: string,
   indexNameAllSitesByDeliveryType: string;
@@ -423,6 +449,7 @@ interface DataAccessConfig {
   pkAllSites: string;
   pkAllLatestAudits: string;
   pkAllOrganizations: string;
+  pkAllConfigurations: string;
 }
 
 export function createDataAccess(