+++ packages/spacecat-shared-data-access/test/unit/service/index.test.js
@@ -59,6 +59,11 @@ describe('Data Access Object Tests', () => {
     'getOrganizationByImsOrgID',
   ];
 
+  const configurationFunctions = [
+    'getConfigurationByVersion',
+    'getConfiguration',
+  ];
+
   let dao;
 
   before(() => {
@@ -88,7 +93,8 @@ describe('Data Access Object Tests', () => {
       ...auditFunctions,
       ...siteFunctions,
       ...siteCandidateFunctions,
-      ...organizationFunctions]);
+      ...organizationFunctions,
+      ...configurationFunctions]);
     Object.keys(dao).forEach((funcName) => {
       expect(expectedFunctions).to.include(funcName);
     });